Move cylc.parsec.LOG to cylc.LOG
