+++ lib/cylc/__init__.py
@@ -19,7 +19,12 @@
 
 import os
 import sys
-from cylc.parsec import LOG
+import logging
+
+
+LOG = logging.getLogger('cylc')  # Acceptable?
+if hasattr(logging, 'NullHandler'):  # Back compat Python <2.7
+    LOG.addHandler(logging.NullHandler())  # Start with a null handler
 
 
 def environ_init():
